requirements.txt
gradio==4.43.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
rank-bm25==0.2.2
transformers==4.44.2
torch>=2.3.0
httpx==0.27.2
tqdm==4.66.5
numpy==1.26.4
scikit-learn==1.5.2
python-dotenv==1.0.1
uvicorn==0.30.6
fastapi==0.112.2
pytest==8.3.2
